requirements.txt
PyYAML==3.12
appdirs==1.4.3